% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{tidy}
\alias{tidy}
\alias{tidy.verbformenobj}
\title{tidy up verbformenobj to tibble}
\usage{
tidy(x, ...)

\method{tidy}{verbformenobj}(obj, ...)
}
\arguments{
\item{x, }{an S3 object of the type "verformenobj"}

\item{..., }{other parameters}
}
\value{
a tibble
}
\description{
This function converts the object generated by verbformen to a tidy tibble.
}
