% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{verbformen}
\alias{verbformen}
\title{Query verbformen}
\usage{
verbformen(input, tidy = FALSE, pos = "ukn", sleep = 1)
}
\arguments{
\item{input}{string, the word to be quried. Please note that verbformen.de accepts only verb, adjective, and noun. Other word types, e.g. adverb, are not supported. It is case-insensitive, i.e. "elefant" and "Elefant" are going to generate the same result.}

\item{tidy}{logical, whether to return a tidy tibble only}

\item{pos}{string, query for which pos of the input, possible values are "verb" (verb), "adj" (adjective), "sub" (noun) and "ukn" (unknown). This is useful for \code{input} that can be functioned as multiple POSs, e.g. "(r)adikal" can be both noun and adjective.}

\item{sleep}{numeric, sleep time after query. Please set it at least to 1 to prevent abuse.}
}
\value{
an S3 object of the type of 'verbformenobj' or a tibble, depending on the parameter \code{tidy}
}
\description{
This function queries the website verbformen.de for information of a word.
}
\author{
Chung-hong Chan \href{mailto:chainsawtiney@gmail.com}{chainsawtiney@gmail.com}
}
